Update channel opening message, Add confirmation on channel deletion by reaction

+++ commands/konsultasi.py
@@ -116,7 +116,7 @@ Anda dapat melakukan konsultasi selama text channel ini ada. Jika Anda sudah sel
 Terimakasih"""
 			embed = discord.Embed(title=f'Halo {payload.member.nick}!', description=desc_message, color=discord.Colour.gold())
 
-			msg = await ch.send(embed=embed)
+			msg = await ch.send(content=message.author.mention, embed=embed)
 			await msg.add_reaction('🔒')
 
 			await message.remove_reaction('🙋', user)
@@ -137,13 +137,35 @@ Terimakasih"""
 							msg.embeds[0].title == f'Halo {payload.member.nick}!')
 
 					if str(payload.emoji) == '🔒' and not(payload.member.bot) and check_message(message):
-						await message.channel.send('Terima kasih telah melakukan konsultasi dengan CTO HMIF ITB!')
-						msg = await message.channel.send('Menghapus channel dalam 3...')
-						await asyncio.sleep(1)
-						for i in range(2, 0, -1):
-							await msg.edit(content=f'Menghapus channel dalam {i}...')
+						await message.channel.send(f'{message.author.mention}\nSebelum menutup channel ini, mohon isi link feedback berikut : https://bit.ly/FeedbackIITConsultation')
+						confirmation = await message.channel.send('Apakah Anda ingin menutup channel ini sekarang?')
+						await confirmation.add_reaction('✅')
+						await confirmation.add_reaction('❌')
+
+						def check(reaction, user):
+							return user == message.author and (str(reaction.emoji) == '✅' or str(reaction.emoji) == '❌')
+
+						try:
+							reaction, _ = await self.bot.wait_for('reaction_add', timeout=60.0, check=check)
+						except asyncio.TimeoutError:
+							await confirmation.clear_reactions()
+							return
+
+						if str(reaction.emoji) == '✅':
+							await confirmation.edit(content='Terima kasih telah melakukan konsultasi dengan CTO HMIF ITB!')
+							await confirmation.clear_reactions()
+
+							msg = await message.channel.send('Menghapus channel dalam 3...')
 							await asyncio.sleep(1)
-						await message.channel.delete()
+							for i in range(2, 0, -1):
+								await msg.edit(content=f'Menghapus channel dalam {i}...')
+								await asyncio.sleep(1)
+							await message.channel.delete()
+
+						elif str(reaction.emoji) == '❌':
+							await confirmation.edit(content='Dibatalkan!', delete_after=5)
+							await confirmation.clear_reactions()
+
 			except KeyError:
 				pass
 	
@@ -153,7 +175,7 @@ Terimakasih"""
 			try:
 				if any(str(message.channel.id) == str(l[-1]) for l in list(db['konsultasi'].values())):
 					if not(message.author.bot):
-						await message.channel.send('Sebelum menutup channel ini, mohon isi link feedback berikut : https://bit.ly/FeedbackIITConsultation')
+						await message.channel.send(f'{message.author.mention}\nSebelum menutup channel ini, mohon isi link feedback berikut : https://bit.ly/FeedbackIITConsultation')
 						confirmation = await message.channel.send('Apakah Anda ingin menutup channel ini sekarang?')
 						await confirmation.add_reaction('✅')
 						await confirmation.add_reaction('❌')
